% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_bigdata.R
\name{cv_bytes_human}
\alias{cv_bytes_human}
\title{Round a byte count to a short human string.}
\usage{
cv_bytes_human(bytes)
}
\description{
Round a byte count to a short human string.
}
\keyword{internal}
