% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_tools_core.R
\docType{data}
\name{CV_CLUSTO_DEFAULT_COLS}
\alias{CV_CLUSTO_DEFAULT_COLS}
\title{Column names `addClustoData` writes by default.}
\format{
An object of class \code{character} of length 2.
}
\usage{
CV_CLUSTO_DEFAULT_COLS
}
\description{
Column names `addClustoData` writes by default.
}
\keyword{internal}
