% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_object_store.R
\name{cv_describe_cellset}
\alias{cv_describe_cellset}
\title{Describe a CellSet (a reusable set of cell barcodes from a cluster/subset)}
\usage{
cv_describe_cellset(x)
}
\description{
Describe a CellSet (a reusable set of cell barcodes from a cluster/subset)
}
\keyword{internal}
