% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_object_store.R
\name{cv_handle_prefix}
\alias{cv_handle_prefix}
\title{Choose a handle prefix based on object class}
\usage{
cv_handle_prefix(value)
}
\description{
Choose a handle prefix based on object class
}
\keyword{internal}
