% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_bigdata.R
\docType{data}
\name{CV_MEMORY_BUDGET_FRACTION}
\alias{CV_MEMORY_BUDGET_FRACTION}
\title{Fraction of total RAM a single R process may reasonably plan to use.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CV_MEMORY_BUDGET_FRACTION
}
\description{
Fraction of total RAM a single R process may reasonably plan to use.
}
\keyword{internal}
