% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_object_store.R
\name{cv_object_get}
\alias{cv_object_get}
\title{Retrieve the raw object for a handle (server-side use only)}
\usage{
cv_object_get(store, handle)
}
\description{
Retrieve the raw object for a handle (server-side use only)
}
\keyword{internal}
