% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_nlu_routing.R
\name{cv_out_of_scope_text}
\alias{cv_out_of_scope_text}
\title{The user-facing reply for an out-of-scope request.}
\usage{
cv_out_of_scope_text(hit)
}
\description{
The user-facing reply for an out-of-scope request.
}
\keyword{internal}
