% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_prompts.R
\docType{data}
\name{CV_PROMPT_DEFAULT_CATEGORY}
\alias{CV_PROMPT_DEFAULT_CATEGORY}
\title{The category a prompt gets when the user does not choose one.}
\format{
An object of class \code{character} of length 1.
}
\usage{
CV_PROMPT_DEFAULT_CATEGORY
}
\description{
The category a prompt gets when the user does not choose one.
}
\keyword{internal}
