% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_tools_cellmarkup.R
\name{cv_register_cellmarkup_tools}
\alias{cv_register_cellmarkup_tools}
\title{Register the ceLLMarkup annotation tool(s).}
\usage{
cv_register_cellmarkup_tools()
}
\description{
Register the ceLLMarkup annotation tool(s).
}
\keyword{internal}
