% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_tools_core.R
\name{cv_register_core_tools}
\alias{cv_register_core_tools}
\title{Register the core CelliVerse tools}
\usage{
cv_register_core_tools()
}
\description{
Register the core CelliVerse tools
}
\keyword{internal}
