% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_registry.R
\name{cv_registry}
\alias{cv_registry}
\title{Get (and lazily build/cache) the registry}
\usage{
cv_registry(rebuild = FALSE)
}
\description{
Get (and lazily build/cache) the registry
}
\keyword{internal}
