% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_registry.R
\name{cv_registry_metadata}
\alias{cv_registry_metadata}
\title{Registry metadata for the client (Package Browser / Tool Inspector)}
\usage{
cv_registry_metadata(reg = cv_registry())
}
\description{
Registry metadata for the client (Package Browser / Tool Inspector)
}
\keyword{internal}
