% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_llm.R
\name{cv_response_has_tools}
\alias{cv_response_has_tools}
\title{Does a normalized response request tool calls?}
\usage{
cv_response_has_tools(resp)
}
\description{
Does a normalized response request tool calls?
}
\keyword{internal}
