% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceLLMarkup.R
\name{.cv_cellmarkup_candidates_df}
\alias{.cv_cellmarkup_candidates_df}
\title{Convert candidate dicts to a ranked data.frame.}
\usage{
.cv_cellmarkup_candidates_df(cands, top_k)
}
\description{
Convert candidate dicts to a ranked data.frame.
}
\keyword{internal}
