% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_tools_core.R
\name{.cv_note_heavy_cost}
\alias{.cv_note_heavy_cost}
\title{Announce the scale of a heavy run before it starts.}
\usage{
.cv_note_heavy_cost(store, args, tool, warnings)
}
\description{
Announce the scale of a heavy run before it starts.
}
\keyword{internal}
