% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.sparse.R
\name{jaccard.sparse}
\alias{jaccard.sparse}
\title{Compute Jaccard similarity for sparse matrices}
\usage{
jaccard.sparse(mat, num_threads = -1)
}
\arguments{
\item{mat}{A sparse matrix of class \code{Matrix}.}

\item{num_threads}{Integer; number of threads to use. \code{-1} uses all available cores.}
}
\value{
A numeric matrix of Jaccard similarity scores.
}
\description{
Computes pairwise Jaccard similarity between columns of a sparse matrix.
}
\examples{
utils::data("pbmc_small", package = "SeuratObject")

mat <- SeuratObject::LayerData(
  pbmc_small,
  assay = "RNA",
  layer = "counts"
)

binary_mat <- mat
binary_mat@x[] <- 1

jac <- jaccard.sparse(
  binary_mat,
  num_threads = 1
)

}
\seealso{
\code{\link{mutual.rank}}
}
