% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addClustoData.R
\name{addClustoData}
\alias{addClustoData}
\title{Add ClustoCell cluster annotations to a Seurat or SingleCellExperiment object}
\usage{
addClustoData(
  obj,
  clustoCell,
  add_major_clusters = TRUE,
  add_sub_clusters = TRUE,
  major_cluster_name = "ClustoCell_Clusters",
  sub_cluster_name = "ClustoCell_SubClusters"
)
}
\arguments{
\item{obj}{An object of class \code{Seurat} or \code{SingleCellExperiment}.}

\item{clustoCell}{An object of class \code{ClustoCell}, generated via \code{clustoCell()} or
\code{markoClust()}.}

\item{add_major_clusters}{Logical; whether to add major cluster labels to the metadata of \code{obj}.}

\item{add_sub_clusters}{Logical; whether to add sub-cluster labels to the metadata of \code{obj}.}

\item{major_cluster_name}{Character; name of the metadata column to store major cluster labels.}

\item{sub_cluster_name}{Character; name of the metadata column to store sub-cluster labels.}
}
\value{
The input object \code{obj} with additional metadata columns containing
ClustoCell cluster annotations.
}
\description{
Adds major cluster and/or sub-cluster labels stored in a \code{ClustoCell}
object to the cell-level metadata of a Seurat or SingleCellExperiment object.
}
\details{
This function transfers clustering results obtained using \code{clustoCell()}
or \code{markoClust()} into an existing single-cell object by appending
cluster labels as metadata columns. Major clusters and sub-clusters can be
added independently and assigned custom column names.
}
\examples{
utils::data("pbmc_small", package = "SeuratObject")

pbmc_small_cc <- clustoCell(
  data = pbmc_small,
  identify_subclusters = TRUE,
  num_threads = 1,
  verbose = FALSE
)

pbmc_small <- addClustoData(
  obj = pbmc_small,
  clustoCell = pbmc_small_cc,
  add_major_clusters = TRUE,
  add_sub_clusters = TRUE
)

}
\seealso{
\code{\link{clustoCell}}, \code{\link{markoClust}}
}
