% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatasetMarkers.R
\name{getDatasetMarkers}
\alias{getDatasetMarkers}
\title{Collect marker genes from a ClustoCell object}
\usage{
getDatasetMarkers(
  obj,
  clusters = TRUE,
  sub_clusters = TRUE,
  positive_markers = TRUE,
  negative_markers = FALSE,
  medium_markers = FALSE,
  thresh_mode = c("n", "rank"),
  pos_thresh = 25,
  neg_thresh = 20,
  med_thresh = 10,
  verbose = TRUE
)
}
\arguments{
\item{obj}{An object of class \code{ClustoCell}.}

\item{clusters}{Logical; whether to collect markers from major clusters.}

\item{sub_clusters}{Logical; whether to collect markers from sub-clusters.}

\item{positive_markers}{Logical; whether to collect positive markers.}

\item{negative_markers}{Logical; whether to collect negative markers.}

\item{medium_markers}{Logical; whether to collect medium markers.}

\item{thresh_mode}{Character; marker selection strategy. One of:
\itemize{
  \item \code{"rank"}: include all markers with ranks up to the threshold.
  \item \code{"n"}: include only the top n markers (rows) in rank order..
}}

\item{pos_thresh}{Integer; threshold for positive markers.}

\item{neg_thresh}{Integer; threshold for negative markers.}

\item{med_thresh}{Integer; threshold for medium markers.}

\item{verbose}{Logical; whether to display progress messages.}
}
\value{
An object of class \code{DatasetMarkers}.
}
\description{
Extracts positive, negative, and/or medium markers from major clusters and
sub-clusters stored in a \code{ClustoCell} object.
}
\details{
Marker selection can be controlled using rank-based or fixed-size thresholds.
Separate thresholds are applied for positive, negative, and medium markers.
}
\examples{
utils::data("pbmc_small", package = "SeuratObject")

cc <- clustoCell(
  data = pbmc_small,
  identify_subclusters = FALSE,
  num_threads = 1,
  verbose = FALSE
)

markers <- getDatasetMarkers(
  obj = cc,
  sub_clusters = FALSE,
  pos_thresh = 20,
  verbose = FALSE
)

}
\seealso{
\code{\link{clustoCell}}, \code{\link{markoClust}}
}
