% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual.rank.R
\name{mutual.rank}
\alias{mutual.rank}
\title{Compute mutual rank from a similarity matrix}
\usage{
mutual.rank(mat, num_threads = -1)
}
\arguments{
\item{mat}{Numeric symmetric matrix (dense or sparse) representing pairwise similarities
or dissimilarities.}

\item{num_threads}{Integer; number of threads to use. Default \code{-1} uses all available cores.}
}
\value{
A numeric matrix of mutual ranks with the same dimensions as \code{mat}.
}
\description{
Computes the mutual rank (MR) for all pairs in a symmetric similarity
or dissimilarity matrix, a robust measure frequently used to stabilize
pairwise similarity relationships.
}
\examples{
utils::data("pbmc_small", package = "SeuratObject")

mat <- SeuratObject::LayerData(
  pbmc_small,
  assay = "RNA",
  layer = "counts"
)

binary_mat <- sign(mat)

similarity_matrix <- jaccard.sparse(
  binary_mat,
  num_threads = 1
)

mr <- mutual.rank(
  similarity_matrix,
  num_threads = 1
)

}
\seealso{
\code{\link{jaccard.sparse}}, \code{\link{markoClust}}
}
